Iris-Core/requirements.txt
python-telegram-bot>=21.0
flask>=3.0
python-dotenv>=1.0
Pillow>=10.0
icalendar>=5.0
